pyproject.toml
[project]
name = "pmsmanagebot"
version = "0.2.0"
description = "Telegram bot for managing Plex/Emby media servers with WebApp support"
authors = [
    { name = "WithdewHua", email = "i@10101.io" }
]
dependencies = [
    "plexapi==4.13.0",
    "pydantic-settings>=2.7.0",
    "python-telegram-bot==20.0a2",
    "redis>=5.2.1",
    "requests==2.28.0",
    "urllib3==1.26.9",
    "fastapi==0.103.0",
    "uvicorn==0.23.2",
    "python-multipart==0.0.6",
    "itsdangerous>=2.2.0",
]
readme = "README.md"
requires-python = ">= 3.11"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["src/app"]

[dependency-groups]
dev = [
    "fastapi[standard]>=0.103.0",
]
